Frontend refactoring and layer separation
